Refactor location of some utils
