+++ kglib/kgcn_experimental/examples/diagnosis/data.py
@@ -21,7 +21,7 @@ import inspect
 
 from kglib.utils.graph.create.from_queries import build_graph_from_queries
 from kglib.utils.graph.create.query_graph import QueryGraph
-from kglib.utils.iterate import multidigraph_data_iterator
+from kglib.utils.graph.iterate import multidigraph_data_iterator
 
 
 CATEGORICAL_ATTRIBUTES = {'name': ['meningitis', 'flu', 'fever', 'light-sensitivity']}
