+++ kglib/kgcn_experimental/examples/diagnosis/main.py
@@ -24,7 +24,7 @@ from graph_nets.utils_np import graphs_tuple_to_networkxs
 
 from kglib.kgcn_experimental.examples.diagnosis.data import create_concept_graphs, write_predictions_to_grakn, get_all_types, \
     CATEGORICAL_ATTRIBUTES
-from kglib.utils.iterate import multidigraph_node_data_iterator, multidigraph_data_iterator
+from kglib.utils.graph.iterate import multidigraph_node_data_iterator, multidigraph_data_iterator
 from kglib.kgcn_experimental.pipeline.utils import apply_logits_to_graphs, duplicate_edges_in_reverse
 from kglib.kgcn_experimental.network.attribute import CategoricalAttribute, BlankAttribute
 from kglib.kgcn_experimental.network.core import softmax, KGCN
