+++ kglib/kgcn_experimental/network/embedding_test.py
@@ -24,7 +24,7 @@ import tensorflow as tf
 from mock import Mock
 from unittest.mock import patch
 from kglib.kgcn_experimental.network.embedding import common_embedding, attribute_embedding, node_embedding
-from kglib.kgcn_experimental.test.utils import get_call_args
+from kglib.utils.test.utils import get_call_args
 
 
 class TestCommonEmbedding(unittest.TestCase):
