+++ kglib/kgcn_experimental/network/typewise.py
@@ -23,7 +23,7 @@ import sonnet as snt
 import tensorflow as tf
 from mock import Mock
 
-from kglib.kgcn_experimental.test.utils import get_call_args
+from kglib.utils.test.utils import get_call_args
 
 
 class TypewiseEncoder(snt.AbstractModule):
