+++ kglib/kgcn_experimental/network/typewise_test.py
@@ -23,7 +23,7 @@ import numpy as np
 import tensorflow as tf
 from mock import Mock
 
-from kglib.kgcn_experimental.test.utils import get_call_args
+from kglib.utils.test.utils import get_call_args
 from kglib.kgcn_experimental.network.typewise import TypewiseEncoder
