+++ kglib/kgcn_experimental/pipeline/process.py
@@ -23,7 +23,7 @@ import numpy as np
 import tensorflow as tf
 
 from kglib.kgcn_experimental.pipeline.encode import encode_solutions, encode_categorically, augment_data_fields
-from kglib.utils.iterate import multidigraph_node_data_iterator, \
+from kglib.utils.graph.iterate import multidigraph_node_data_iterator, \
     multidigraph_edge_data_iterator, multidigraph_data_iterator
 from kglib.kgcn_experimental.pipeline.feed import create_placeholders, create_feed_dict, make_all_runnable_in_session
 from kglib.kgcn_experimental.network.loss import loss_ops_from_difference
