+++ kglib/utils/graph/iterate.py
